로그인 수정

+++ Subway/WebContent/admin/notice/notice_content.jsp
@@ -78,11 +78,13 @@ div.bor {
 								<a href="http://localhost:9000/Subway/board/notice_list.jsp" onclick="view.list();return false;">목록보기</a>
 							</div>
 							<div class="btn_right">
-								<% if(svo.getEmail().equals("admin@naver.com")){ %>
-								<a href="notice_update.jsp?bid=<%=bid%>&rno=<%=rno%>"><button
-										type="button" class="btn_style2">수정</button></a> <a
-									href="notice_delete.jsp?bid=<%=bid%>&rno=<%=rno%>"><button
-										type="button" class="btn_style2">삭제</button></a>
+								<% if(svo != null) {%>
+									<% if(svo.getEmail().equals("admin@naver.com")){ %>
+										<a href="notice_update.jsp?bid=<%=bid%>&rno=<%=rno%>"><button
+												type="button" class="btn_style2">수정</button></a> <a
+											href="notice_delete.jsp?bid=<%=bid%>&rno=<%=rno%>"><button
+												type="button" class="btn_style2">삭제</button></a>
+									<% } %>
 								<% } %>
 							</div>
 						</div>
